[WDP-20] Przeniesienie view mode do utils
