+++ src/components/layout/MainLayout/MainLayout.js
@@ -5,20 +5,29 @@ import Header from '../Header/Header';
 import Footer from '../Footer/Footer';
 
 import { useDispatch } from 'react-redux';
-import { updateViewMode } from '../../../redux/viewModeRedux';
+import { setViewMode } from '../../../redux/viewModeRedux';
+
+import { getViewMode } from '../../../utils/viewMode';
 
 const MainLayout = ({ children }) => {
   const dispatch = useDispatch();
+  let activeView = getViewMode(window.innerWidth);
 
   useEffect(() => {
-    dispatch(updateViewMode(window.innerWidth));
+    dispatch(setViewMode(activeView));
 
     const handleResize = () => {
-      dispatch(updateViewMode(window.innerWidth));
+      const newView = getViewMode(window.innerWidth);
+      if (activeView !== newView) {
+        dispatch(setViewMode(newView));
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        activeView = newView;
+      }
     };
     window.addEventListener('resize', handleResize);
     return () => window.removeEventListener('resize', handleResize);
-  }, [dispatch]);
+    // eslint-disable-next-line
+  }, []);
 
   return (
     <div>
