+++ src/redux/viewModeRedux.js
@@ -1,20 +1,5 @@
-const viewModes = [
-  {
-    type: 'mobile',
-    max: 400,
-  },
-  {
-    type: 'tablet',
-    max: 750,
-  },
-  {
-    type: 'desktop',
-    max: 2400,
-  },
-];
-
 /* selectors */
-export const getViewMode = ({ activeViewMode }) => activeViewMode;
+export const viewMode = ({ activeViewMode }) => activeViewMode;
 
 /* action name creator */
 const reducerName = 'viewMode';
@@ -25,15 +10,6 @@ const SET_VIEW = createActionName('SET_VIEW');
 
 /* action creators */
 export const setViewMode = payload => ({ payload, type: SET_VIEW });
-
-/* action creator effects */
-export const updateViewMode = width => {
-  return dispatch => {
-    const view = viewModes.find(view => width < view.max);
-    console.log('ufsowa view: ', width, view);
-    dispatch(setViewMode(view ? view.type : 'desktop'));
-  };
-};
 
 /* reducer */
 export default function reducer(statePart = [], action = {}) {
